% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air_quality_leeds}
\alias{air_quality_leeds}
\title{Air quality data at two points in Leeds}
\format{
A data frame with 10 rows and 22 variables:
\describe{
\item{\code{site}}{site name}
\item{\code{code}}{site code}
\item{\code{site_type}}{site type}
\item{\code{geometry}}{list Geometry}
\item{\code{date}}{date}
\item{\code{nox}}{concentration of nox}
\item{\code{no2}}{concentration of no2}
}
}
\usage{
air_quality_leeds
}
\description{
This dataset represents routes associated with the top 10
commuter 'desire lines' in Leeds.
}
\examples{
dim(air_quality_leeds)
names(air_quality_leeds)
table(air_quality_leeds$site)
}
\keyword{datasets}
