% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_td.R
\name{get_td}
\alias{get_td}
\title{Get transport data}
\usage{
get_td(
  file_name,
  base_url = "https://github.com/ITSLeeds/transportdata/releases/download/0.1/",
  full_url = NULL,
  download = !grepl(pattern = "\\\\.Rds", x = add_extension(file_name))
)
}
\arguments{
\item{file_name}{The name of the file to download, with or without extension}

\item{base_url}{Where to download the data from? The \code{transportdata} repo
by default}

\item{full_url}{The full URL of the file to download. Usually not needed.}

\item{download}{Should the file be downloaded to disk? \code{FALSE}
by default unless the file is not an \code{Rds} file.}
}
\description{
Get transport data
}
\examples{
# piggyback::pb_download_url("routes_leeds_foot.Rds")
library(sf)
rlw = get_td(file_name = "routes_leeds_foot")
}
