% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_leeds_foot_10}
\alias{routes_leeds_foot_10}
\alias{routes_leeds_bike_10}
\alias{routes_leeds_car_10}
\title{Routes on the transport network}
\format{
A data frame with 10 rows and 22 variables:
\describe{
\item{\code{distance}}{double Distance in m according to OSRM}
\item{\code{duration}}{double Duration in seconds according to OSRM}
\item{\code{geometry}}{list Geometry}
}
}
\usage{
routes_leeds_foot_10
}
\description{
This dataset represents routes associated with the top 10
commuter 'desire lines' in Leeds.
}
\examples{
plot(routes_leeds_foot_10$geometry, lwd = 6, col = "green")
plot(routes_leeds_bike_10$geometry, lwd = 4, col = "blue", add = TRUE)
plot(routes_leeds_car_10$geometry, lwd = 2, col = "red", add = TRUE)
}
\keyword{datasets}
