% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_survey_leeds}
\alias{time_survey_leeds}
\title{Time use survey data}
\format{
A data frame with 2826 rows and 20 variables.
}
\usage{
time_survey_leeds
}
\description{
See the \href{http://apollochoicemodelling.com/files/apollo_timeUseData_dictionary.pdf}{data dictionary for description}
}
\examples{
head(time_survey_leeds)
}
\keyword{datasets}
