% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_leeds}
\alias{zones_leeds}
\title{Zones of administrative (MSOA) zones in Leeds, UK}
\format{
A spatial (sf) data frame with 107 rows and 15 variables (modes same as \code{od_leeds} dataset).
\describe{
\item{\code{geo_code}}{character geo_code of the unique zones}
\item{\code{dutch_slc}}{Number who cycle to work under the PCT's Go Dutch scenario}
}
}
\usage{
zones_leeds
}
\description{
This dataset contains polygons and Census travel to work data associated with
administrative (MSOA) zones in Leeds, UK
}
\examples{
zones_leeds
dim(zones_leeds)
library(sf)
plot(zones_leeds)
}
\keyword{datasets}
